Launching issues
- fixes #163 cannot launch from activity stream
- fixes #131 vimeo home context menu support
- fixes #111 vimeo detail context menu support
- factor context menu handling into `/lib/context-menu-handlers.js`
- move `sendMetricsData` and `launchVideo` into their own files in `lib`
